refactor: 인증/인가 예외 처리 단일화 및 패키지 구조 정리
