+++ src/main/java/com/sprint/mission/discodeit/security/handler/ForbiddenAccessDeniedHandler.java
@@ -1,12 +1,12 @@
-package com.sprint.mission.discodeit.security;
+package com.sprint.mission.discodeit.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sprint.mission.discodeit.exception.DiscodeitException;
 import com.sprint.mission.discodeit.exception.ErrorResponse;
+import com.sprint.mission.discodeit.security.translator.SecurityExceptionTranslator;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.time.Instant;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.AccessDeniedException;
@@ -18,20 +18,15 @@ import org.springframework.stereotype.Component;
 public class ForbiddenAccessDeniedHandler implements AccessDeniedHandler {
 
   private final ObjectMapper objectMapper;
+  private final SecurityExceptionTranslator translator;
 
   @Override
   public void handle(HttpServletRequest request, HttpServletResponse response,
       AccessDeniedException accessDeniedException) throws IOException {
-    ErrorResponse errorResponse = new ErrorResponse(
-        "FORBIDDEN",
-        "접근 권한이 없습니다.",
-        accessDeniedException.getClass().getSimpleName(),
-        HttpServletResponse.SC_FORBIDDEN,
-        Instant.now(),
-        Map.of()
-    );
+    DiscodeitException discodeitException = translator.translate(accessDeniedException);
+    ErrorResponse errorResponse = ErrorResponse.from(discodeitException);
 
-    response.setStatus(HttpServletResponse.SC_FORBIDDEN);
+    response.setStatus(discodeitException.getErrorCode().getStatus());
     response.setContentType(MediaType.APPLICATION_JSON_VALUE);
     objectMapper.writeValue(response.getWriter(), errorResponse);
   }
