+++ src/main/java/com/sprint/mission/discodeit/security/handler/LoginFailureHandler.java
@@ -1,14 +1,13 @@
-package com.sprint.mission.discodeit.security;
+package com.sprint.mission.discodeit.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sprint.mission.discodeit.exception.ErrorCode;
+import com.sprint.mission.discodeit.exception.DiscodeitException;
 import com.sprint.mission.discodeit.exception.ErrorResponse;
+import com.sprint.mission.discodeit.security.translator.SecurityExceptionTranslator;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.time.Instant;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
@@ -20,21 +19,15 @@ import org.springframework.stereotype.Component;
 public class LoginFailureHandler implements AuthenticationFailureHandler {
 
   private final ObjectMapper objectMapper;
+  private final SecurityExceptionTranslator translator;
 
   @Override
   public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
       AuthenticationException exception) throws IOException, ServletException {
+    DiscodeitException discodeitException = translator.translate(exception);
+    ErrorResponse errorResponse = ErrorResponse.from(discodeitException);
 
-    ErrorResponse errorResponse = new ErrorResponse(
-        ErrorCode.UNAUTHORIZED.name(),
-        ErrorCode.UNAUTHORIZED.getMessage(),
-        exception.getClass().getSimpleName(),
-        HttpServletResponse.SC_UNAUTHORIZED,
-        Instant.now(),
-        Map.of()
-    );
-
-    response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+    response.setStatus(discodeitException.getErrorCode().getStatus());
     response.setContentType(MediaType.APPLICATION_JSON_VALUE);
     objectMapper.writeValue(response.getWriter(), errorResponse);
   }
