+++ src/main/java/com/sprint/mission/discodeit/security/userdetails/DiscodeitUserDetails.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.security;
+package com.sprint.mission.discodeit.security.userdetails;
 
 import com.sprint.mission.discodeit.dto.response.UserResponse;
 import java.util.Collection;
